% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_generated.R
\docType{data}
\name{precip_generated}
\alias{precip_generated}
\title{Water storage function}
\format{A data frame with 5 rows and 1 column
\describe{
\item{precip}-{Generated precipitation values in inches per year}
}}
\usage{
precip_generated
}
\description{
Generated data
}
\keyword{datasets}
