% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterstorage.R
\name{waterstorage}
\alias{waterstorage}
\title{Water storage function}
\usage{
waterstorage(precip, evap, runoff, watershed_size)
}
\arguments{
\item{precip}{a data frame that has rainfall values in inches per year}

\item{evap}{a data frame that has evapotranspiration values in inches per year}

\item{runoff}{a data frame that has runoff values in inches per year}

\item{watershed_size}{a value in square miles}
}
\value{
storage the water stored in a watershed in a year, units in cubic feet per year
}
\description{
Function describing water stored in a watershed/year using simplified inflow/outflow processes
Function assumptions: The user must provide watershed_size in square miles and precipitation, evaporation, and runoff in inches per year
}
